test(runtime): 独立保留维护版跨平台写入回归测试
